api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
aiofiles==24.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
mysql-connector-python==9.1.0
python-dotenv==1.0.1
pydantic==2.10.4
python-multipart==0.0.20
